pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectra-notes"
version = "0.1.0"
description = "FastAPI + PostgreSQL(pgvector) semantic search API"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "python-jose[cryptography]>=3.3",
  "sentence-transformers>=3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E","F","I","UP"]
